rewrite logger as typescript
