--- src/lib/logger.js
@@ -1,15 +1,0 @@
-const winston = require('winston');
-const logger = winston.createLogger({
-    level: 'info',
-    format: winston.format.json(),
-    defaultMeta: { service: 'user-service' },
-    transports: [
-        new winston.transports.Console(),
-        new winston.transports.File({ filename: 'combined.log' })
-    ],
-    exceptionHandlers: [
-        new winston.transports.File({ filename: 'exceptions.log' })
-    ]
-});
-
-module.exports = logger;
